pref: 删除时修改主页状态 & 重新生成 mapping.json
