+++ scripts/start-all.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-npm run generate & 
 npm run dev &
 npm run server &
 wait
