--- src/layout/122a14c4ab646d5a612d74ceaddd20a0.json
@@ -1,1 +1,0 @@
-{"BlockGroup":{"block-623187":{"x":158,"y":256,"style":{"width":100,"height":80,"background":"#F96","borderRadius":"10px","border":"1px solid #aaa"}},"block-624018":{"x":367,"y":368,"style":{"width":100,"height":80,"background":"#F96","borderRadius":"10px","border":"1px solid #aaa"}},"block-73377":{"x":253,"y":525,"style":{"width":100,"height":80,"background":"#F96","borderRadius":"10px","border":"1px solid #aaa"}},"block-66174":{"x":80,"y":657,"style":{"width":100,"height":80,"background":"#F96","borderRadius":"10px","border":"1px solid #aaa"}},"block-74386":{"x":426,"y":477,"style":{"width":100,"height":80,"background":"#F96","borderRadius":"10px","border":"1px solid #aaa"}}},"TagGroup":{"tag-626505":{"x":167,"y":284,"style":{"width":100,"height":32},"editable":false,"input":"test"},"tag-629962":{"x":405,"y":398,"style":{"width":100,"height":32},"editable":false,"input":"test2"},"tag-80986":{"x":286,"y":555,"style":{"width":100,"height":32},"editable":false,"input":"test3"},"tag-67933":{"x":110,"y":689,"style":{"width":100,"height":32},"editable":false,"input":"测试"},"tag-76032":{"x":438,"y":508,"style":{"width":100,"height":32},"editable":false,"input":"托尔斯泰"}}}
