+++ src/mapping.json
@@ -1,1 +1,1 @@
-[{"id":"122a14c4ab646d5a612d74ceaddd20a0","title":"test1","thumbnailUrl":"dist/assets/5795774_0.jpg","hoverText":"测试 layout-122a14c4ab646d5a612d74ceaddd20a0.json","createTime":1548817524632,"modifyTime":1548817524632},{"id":"1a79a4d60de6718e8e5b326e338ae533","title":"test1","thumbnailUrl":"dist/assets/5795774_0.jpg","hoverText":"测试 layout-1a79a4d60de6718e8e5b326e338ae533.json","createTime":1548817524632,"modifyTime":1548817524632}]
+[]
